Move templates back
