+++ lib/invitations/emailTemplates/inviteToCollaborate.ts
