+++ lib/invitations/emailTemplates/ownerRemovedNotification.ts
