+++ lib/invitations/emailTemplates/youHaveBeenRemovedNotification.ts
