+++ lib/invitations/inviteUserByEmail.ts
@@ -5,8 +5,8 @@ import { getTemplateWithAssociatedUsers } from "@lib/templates";
 import { AccessControlError } from "@lib/privileges";
 import { prisma } from "@lib/integration/prismaConnector";
 import { sendEmail } from "@lib/integration/notifyConnector";
-import { inviteToCollaborate } from "@lib/emailTemplates/inviteToCollaborate";
-import { inviteToForms } from "@lib/emailTemplates/inviteToForms";
+import { inviteToCollaborate } from "@lib/invitations/emailTemplates/inviteToCollaborate";
+import { inviteToForms } from "@lib/invitations/emailTemplates/inviteToForms";
 import { getOrigin } from "@lib/origin";
 import { logMessage } from "@lib/logger";
 import { Invitation } from "@prisma/client";
